+++ .stylelintrc-css.mjs
@@ -4,5 +4,6 @@ export default {
   extends: "stylelint-config-standard",
   rules: {
     "selector-class-pattern": null,
+    "color-function-notation": ["modern", { ignore: ["with-var-inside"] }],
   },
 }